requirements.txt
flask>=2.0.0
transformers>=4.30.0
torch>=2.0.0
pyyaml>=6.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protocol-buffers>=4.21.0
